refactor: rename console metric exporter
